requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-docx
pandas
python-dateutil
pydantic